pyproject.toml
[project]
name = "shopware-bmecat-importer"
version = "0.1.0"
description = "BMEcat ETIM importer for Shopware 6 using Sync API"
authors = [ { name = "Shopware Import Bot" } ]
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "lxml>=5.1",
    "PyYAML>=6.0",
    "pydantic>=2.6",
    "python-dateutil>=2.8",
]
readme = "README.md"

[project.scripts]
shopware-bmecat = "shopware_bmecat.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
minversion = "7.4"
addopts = "-q"
testpaths = ["tests", "Tests"]
